backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
web3==6.15.1
eth-account==0.10.0